+++ src/Resources/app/storefront/src/sg-webcheckout-app/event.manager.js
@@ -1,6 +1,7 @@
 import CloseBrowserEvent from './events/closeBrowser.event';
 import LoginEvent from './events/login.event';
 import PurchaseEvent from './events/purchase.event';
+import TokenSyncEvent from './events/tokenSync.event';
 
 export default class SGWebcheckoutEventManager {
     /**
@@ -24,6 +25,7 @@ export default class SGWebcheckoutEventManager {
         this.registerEvent(CloseBrowserEvent);
         this.registerEvent(LoginEvent);
         this.registerEvent(PurchaseEvent);
+        this.registerEvent(TokenSyncEvent);
     }
 
     /**
@@ -38,7 +40,7 @@ export default class SGWebcheckoutEventManager {
             if (!event.supports(this.controllerName, this.actionName, this.properties) || !event.active) {
                 return;
             }
-            event.log('Executing event > ' + event.constructor.name);
+            event.log('Executing event > ' + event.constructor.name); // works on non-minified
             event.execute(this.properties);
         });
     }