+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+# 2.2.2
+
+- fixed inApp session drops after visiting checkout/profile pages for SW 6.5.8.9+, 6.6.1.1+
+- fixed issue with double JS load on storefront
+- removed compiled frontend JS file due to lack of compatibility between 6.5 & 6.6
+
 # 2.2.1
 
 - fixes for admin JS & vue3